+++ node/lib/DataConsumer.js
@@ -181,7 +181,7 @@ class DataConsumer extends EnhancedEventEmitter_1.EnhancedEventEmitter {
             message = ' ';
         else if (ppid === 57)
             message = Buffer.alloc(1);
-        const requestData = { ppid };
+        const requestData = String(ppid);
         await this.#payloadChannel.request('dataConsumer.send', this.#internal.dataConsumerId, requestData, message);
     }
     /**