+++ node/lib/DataProducer.js
@@ -167,7 +167,7 @@ class DataProducer extends EnhancedEventEmitter_1.EnhancedEventEmitter {
             message = ' ';
         else if (ppid === 57)
             message = Buffer.alloc(1);
-        const notifData = { ppid };
+        const notifData = String(ppid);
         this.#payloadChannel.notify('dataProducer.send', this.#internal.dataProducerId, notifData, message);
     }
     handleWorkerNotifications() {