+++ worker/include/Utils.hpp
@@ -7,6 +7,7 @@
 #include <cstring> // std::memcmp(), std::memcpy()
 #include <nlohmann/json.hpp>
 #include <string>
+#include <vector>
 #ifdef _WIN32
 #include <ws2ipdef.h>
 // https://stackoverflow.com/a/24550632/2085408
@@ -276,6 +277,8 @@ namespace Utils
 		static uint8_t* Base64Decode(const uint8_t* data, size_t len, size_t& outLen);
 
 		static uint8_t* Base64Decode(const std::string& str, size_t& outLen);
+
+		static std::vector<std::string> Split(const std::string& str, char separator, size_t limit = 0);
 	};
 
 	class Time