+++ worker/src/RTC/DataConsumer.cpp
@@ -8,6 +8,7 @@
 #include "Utils.hpp"
 #include "Channel/ChannelNotifier.hpp"
 #include "RTC/SctpAssociation.hpp"
+#include <stdexcept>
 
 namespace RTC
 {
@@ -198,12 +199,10 @@ namespace RTC
 				// Trigger 'bufferedamountlow' now.
 				if (this->bufferedAmount <= this->bufferedAmountLowThreshold)
 				{
-					// Notify the Node DataConsumer.
-					json data = json::object();
+					std::string data("{\"bufferedAmount\":\"");
 
-					data["bufferedAmount"] = this->bufferedAmount;
-
-					Channel::ChannelNotifier::Emit(this->id, "bufferedamountlow", data);
+					data.append(std::to_string(this->bufferedAmount));
+					data.append("\"}");
 				}
 				// Force the trigger of 'bufferedamountlow' once there is less or same
 				// buffered data than the given threshold.
@@ -240,26 +239,29 @@ namespace RTC
 					MS_THROW_ERROR("no SCTP association present");
 				}
 
-				auto jsonPpidIt = request->data.find("ppid");
+				int ppid;
 
-				if (jsonPpidIt == request->data.end() || !Utils::Json::IsPositiveInteger(*jsonPpidIt))
+				// This may throw.
+				// NOTE: If this throws we have to catch the error and throw a MediaSoupError
+				// intead, otherwise the process would crash.
+				try
 				{
-					MS_THROW_TYPE_ERROR("invalid ppid");
+					ppid = std::stoi(request->data);
+				}
+				catch (const std::exception& error)
+				{
+					MS_THROW_TYPE_ERROR("invalid PPID value: %s", error.what());
 				}
 
-				auto ppid       = jsonPpidIt->get<uint32_t>();
 				const auto* msg = request->payload;
 				auto len        = request->payloadLen;
 
 				if (len > this->maxMessageSize)
 				{
-					MS_WARN_TAG(
-					  message,
+					MS_THROW_TYPE_ERROR(
 					  "given message exceeds maxMessageSize value [maxMessageSize:%zu, len:%zu]",
 					  len,
 					  this->maxMessageSize);
-
-					return;
 				}
 
 				const auto* cb = new onQueuedCallback(
@@ -338,9 +340,10 @@ namespace RTC
 			this->forceTriggerBufferedAmountLow = false;
 
 			// Notify the Node DataConsumer.
-			json data = json::object();
+			std::string data("{\"bufferedAmount\":\"");
 
-			data["bufferedAmount"] = this->bufferedAmount;
+			data.append(std::to_string(this->bufferedAmount));
+			data.append("\"}");
 
 			Channel::ChannelNotifier::Emit(this->id, "bufferedamountlow", data);
 		}