+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * Node: Define TypeScript types for `internal` and `data` objects (PR #891).
 * `Channel` and `PayloadChannel`: Refactor `internal` with a single `handlerId` (PR #889).
+* `Channel` and `PayloadChannel`: Optimize message format and JSON generation (PR #893).
 * Update NPM deps.